Do not rely on the SN_EXISTS bit flag to detect empty values.
This patch frees the SN_EXISTS bit from the storage_number format. We detect
empty values with a previously unused representation of the storage_number,
ie. 0 with a non-empty exponent equal to 001.

The bit flag has been renamed to SN_ANOMALOUS. It will be used in
follow-up patches to track the anomaly status of the value represented by
a storage_number.