Do not rely on the SN_EXISTS bit flag to detect empty values.
This patch frees the SN_EXISTS bit. We detect empty values with
a previously unused representation of the storage number, ie. 0
with a non-empty exponent equal to 001.

Fix exporting tests

The numbers are different because the SN_EXISTS bit is not set anymore.

Memset DB pages to SN_EMPTY_SLOT.