Block replication thread when we are also receiving gaps.
For the time being, we block the replication thread regardless of
whether the sender gap overlaps with the receiver gaps or not. We can
optimize this once we verify that active-active setup works correctly.

Another thing to point out is that based on the current implementation,
it is possible to skip sending upstream the gap data for a single dimension,
ie. up to 1024 seconds, in the worst case scenario. For this to happen:

    - The sender/receiver gap has to overlap,
    - The child has to connect in between the time it takes to
      check the size of the receiver's gaps and perform the query
      for the dimension the replication thread is currently filling.

Considering how straightforward the current implementation is, and the
fact that the data loss is bounded to just one dimension and less than
1024 seconds, it should be fine for the time being.